PyNexRad/requirements.txt
boto3
botocore
matplotlib
metpy
# mpl_toolkits
numpy
pyimgur
pykafka
pytest
pickle